update code to fix issues from pylint
